fix: rename rule to avoid only unsupported SSR methods
