+++ docs/rules/no-unsupported-ssr-properties.md
@@ -1,4 +1,4 @@
-# Disallow access of properties on this during SSR (`lwc/no-this-property-during-ssr`)
+# Disallow access of properties on this during SSR (`lwc/no-unsupported-ssr-properties`)
 
 Browser APIs must not be accessed when SSR is being done. This rule prevents usage of browser APIs like `querySelector`, `template`
 on `this` in `connectedCallback` (and in methods called from `connectedCallback` or anywhere when SSR is being done).
