+++ lib/index.js
@@ -29,7 +29,7 @@ const rules = {
     'valid-track': require('./rules/valid-track'),
     'valid-wire': require('./rules/valid-wire'),
     'no-restricted-browser-globals-during-ssr': require('./rules/no-restricted-browser-globals-during-ssr'),
-    'no-this-property-during-ssr': require('./rules/no-this-property-during-ssr'),
+    'no-unsupported-ssr-properties': require('./rules/no-unsupported-ssr-properties'),
 };
 
 module.exports = {
