+++ lib/rules/no-unsupported-ssr-properties.js
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2020, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+'use strict';
+const { noPropertyAccessDuringSSR } = require('../rule-helpers');
+const { docUrl } = require('../util/doc-url');
+
+const disallowedProperties = [
+    'dispatchEvent',
+    'firstChild',
+    'firstElementChild',
+    'getBoundingClientRect',
+    'getElementsByClassName',
+    'getElementsByTagName',
+    'lastChild',
+    'lastElementChild',
+    'ownerDocument',
+    'querySelector',
+    'querySelectorAll',
+    'removeEventListener',
+];
+module.exports = {
+    meta: {
+        type: 'problem',
+        docs: {
+            url: docUrl('no-unsupported-ssr-properties'),
+            category: 'LWC',
+            description: 'disallow access of unsupported properties in SSR',
+        },
+        schema: [],
+        messages: {
+            propertyAccessFound: '`{{ identifier }}` is unsupported in SSR.',
+        },
+    },
+    create: (context) => {
+        return noPropertyAccessDuringSSR(disallowedProperties, (node) => {
+            context.report({
+                node,
+                messageId: 'propertyAccessFound',
+                data: {
+                    identifier: node.property.name,
+                },
+            });
+        });
+    },
+};
